rename variable 'visible' to 'iconsVisible', add tag form, put active-messenger in its own folder
