+++ nginx.conf
@@ -2,9 +2,9 @@ server {
     location / {
 	add_header 'Access-Control-Allow-Origin' '*';
         root /usr/share/nginx/html;
-        index index.html;
+        index.js index.js.html;
         autoindex on;
-        set $fallback_file /index.html;
+        set $fallback_file /index.js.html;
         try_files $uri $fallback_file;
     }
