+++ src/components/active-messenger/index.js
@@ -1,0 +1,3 @@
+import { ActiveMessenger } from './active-messenger';
+
+export default ActiveMessenger;
